Create README.md
Fortran 90 programs for analyzing various stuff of nucleosome. The required sample input files are provided in the INPUT folder with the same name as that of the program with an extension .inp. Some programs will be used to analyse dcd trajectories. To run these programs, one requires a reference pdb file that will be basically needed for selection. Rest of the programs will be used either for post-processing Curves+ generated data or a simple program. The user guide is provided at the top of each of the programs. If there is any issue, please directly send an email to my mail id prabir07chem@gmail.com 

N.B. - All the input files given are example input files. Users need to modify the input file as per their need.